Fix NullPointerException in editGroup endpoint
Previously, editGroup would throw a NullPointerException due to the fact
that no user was defined to the group after editing, only their UUID,
which was not provided by the owner upon editing. This has been fixed,
as now the user is added by the server. Technically, this could easily
be circumvented by simply not changing the user, but this method
guarantees consistency.